+++ tests/unit/index.test.js
@@ -10,7 +10,7 @@ describe('index.js Component Tests', () => {
   let index;
 
   beforeAll(async () => {
-    index = await import('../../index.js');
+    index = await import('../../scraper/index.js');
   });
 
   beforeEach(() => {